# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rayi API",
    "version": "0.1.0",
    "description": "Conditional payment rail for brand and creator collaborations. Money crosses this API as integer minor units in a string, never as a JSON number."
  },
  "paths": {
...
    "/v1/me/sessions": {
      "get": {
        "operationId": "listMySessions",
        "summary": "List your own active sessions",
        "tags": [
          "account"
        ],
        "description": "Returns every unexpired session for the signed-in user. Session tokens are never included — a list that returned them would turn \"show me my devices\" into \"hand me a credential for each\".",
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "sessionId": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          },
                          "expiresAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          },
                          "ipAddress": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "userAgent": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "current": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "sessionId",
                          "createdAt",
                          "expiresAt",
                          "ipAddress",
                          "userAgent",
                          "current"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "sessions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-rayi-access": {
          "kind": "self",
          "stepUp": false
        }
      }
    },
    "/v1/me/sessions/{sessionId}": {
      "delete": {
        "operationId": "revokeMySession",
        "summary": "Revoke one of your own sessions",
        "tags": [
          "account"
        ],
        "description": "Returns 404 for a session that is not yours, so the endpoint does not confirm which session ids exist.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {},
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-rayi-access": {
          "kind": "self",
          "stepUp": false
        }
      }
    },
    "/v1/me/sessions/revoke-others": {
      "post": {
        "operationId": "revokeMyOtherSessions",
        "summary": "Sign out everywhere except here",
        "tags": [
          "account"
        ],
        "description": "The \"I think I have been compromised\" action. Deliberately keeps the current session alive: signing a user out of the device they are using, at the moment they are securing their account, forces them back through the same email an attacker may control.",
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "revoked": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "revoked"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-rayi-access": {
          "kind": "self",
          "stepUp": false
        }
      }
    },
    "/v1/me/profile": {
      "patch": {
        "operationId": "updateMyProfile",
        "summary": "Update your own profile",
        "tags": [
          "account"
        ],
        "description": "An explicit allowlist of fields. Better Auth’s /update-user accepts a partial user object, which is how `role`, `twoFactorEnabled` or `isEmailVerified` become writable by anyone holding a session. Email is not here: changing the address that receives magic links is an account-takeover primitive and needs its own flow.",
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "updated": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "updated"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-rayi-access": {
          "kind": "self",
          "stepUp": false
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "lastName": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "bio": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "image": {
                    "type": "string",
                    "maxLength": 2000,
                    "format": "uri"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/v1/me/activity": {
      "get": {
        "operationId": "listMyActivity",
        "summary": "Your own security activity",
        "tags": [
          "account"
        ],
        "description": "The audit events this user caused. Showing them is what makes the audit log useful to the person most likely to notice something wrong first.",
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "occurredAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          },
                          "action": {
                            "type": "string"
                          },
                          "ipAddress": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "occurredAt",
                          "action",
                          "ipAddress"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "events"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-rayi-access": {
          "kind": "self",
          "stepUp": false
        }
      }
    },
    "/v1/me/step-up": {
      "post": {
        "operationId": "startStepUp",
        "summary": "Confirm your identity before a sensitive action",
        "tags": [
          "account"
        ],
        "description": "Verifies a code from your authenticator app and mints a short-lived, single-use grant bound to one purpose. A session cannot prove who is at the keyboard right now — it was established once, possibly days ago, possibly on a device that is no longer in your hands.",
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    }
                  },
                  "required": [
                    "expiresAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-rayi-access": {
          "kind": "self",
          "stepUp": false
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "account:change_email",
                      "account:disable_two_factor",
                      "account:enable_two_factor",
                      "member:change_role",
                      "member:remove",
                      "money_authority:grant",
                      "funds:release"
                    ]
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\s*\\d{3}\\s*\\d{3}\\s*$|^\\d{6}$"
                  }
                },
                "required": [
                  "purpose",
                  "code"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/v1/me/email": {
      "post": {
        "operationId": "requestEmailChange",
        "summary": "Start changing your email address",
        "tags": [
          "account"
        ],
        "description": "Needs a fresh authenticator code, and notifies the OLD address. The address that receives magic links IS the account, so changing it is the highest-value takeover step available — and the person who must hear about it is whoever holds the address today.",
        "responses": {
          "202": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "pending_confirmation"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Problem details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-rayi-access": {
          "kind": "self",
          "stepUp": true
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "newEmail": {
                    "type": "string",
                    "maxLength": 255,
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "code": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 12
                  }
                },
                "required": [
                  "newEmail",
                  "code"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Problem": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI reference identifying the problem type."
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary. Safe to reword — do not branch on it."
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "HTTP status code."
          },
          "detail": {
            "description": "Human-readable explanation specific to this occurrence.",
            "type": "string"
          },
          "instance": {
            "description": "URI reference identifying this specific occurrence.",
            "type": "string"
          },
          "code": {
            "type": "string",
            "enum": [
              "validation_failed",
              "unauthenticated",
              "forbidden",
              "step_up_required",
              "not_found",
              "idempotency_key_reused",
              "insufficient_unallocated_funds",
              "budget_envelope_exceeded",
              "daily_limit_exceeded",
              "deposit_not_settled",
              "organization_frozen",
              "conflict",
              "rate_limited",
              "internal_error"
            ],
            "description": "Stable machine-readable code. THIS is the contract."
          },
          "requestId": {
            "type": "string",
            "description": "Correlates to server logs. Always show this in support surfaces."
          },
          "challenge": {
            "type": "object",
            "properties": {
              "challengeId": {
                "type": "string"
              },
              "reason": {
                "type": "string"
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
              }
            },
            "required": [
              "challengeId",
              "reason",
              "expiresAt"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "requestId"
        ],
        "additionalProperties": false,
        "description": "RFC 9457 problem details."
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-rayi.session",
        "description": "Host-only session cookie. The API is served same-origin with the app, so there is no CORS, no preflight and no SameSite=None."
      }
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ]
}